+++ alert-common/src/main/java/com/synopsys/integration/alert/common/provider/ProviderValidator.java
@@ -1,0 +1,6 @@
+package com.synopsys.integration.alert.common.provider;
+
+public abstract class ProviderValidator {
+
+    public abstract boolean validate();
+}
