+++ src/main/java/com/synopsys/integration/alert/workflow/startup/component/SystemValidator.java
@@ -22,13 +22,11 @@
  */
 package com.synopsys.integration.alert.workflow.startup.component;
 
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import org.apache.commons.lang3.BooleanUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -37,29 +35,21 @@ import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Component;
 
 import com.synopsys.integration.alert.ProxyManager;
-import com.synopsys.integration.alert.common.AlertProperties;
 import com.synopsys.integration.alert.common.enumeration.SystemMessageSeverity;
 import com.synopsys.integration.alert.common.enumeration.SystemMessageType;
-import com.synopsys.integration.alert.common.exception.AlertRuntimeException;
 import com.synopsys.integration.alert.common.persistence.model.UserModel;
+import com.synopsys.integration.alert.common.provider.ProviderValidator;
 import com.synopsys.integration.alert.common.security.EncryptionUtility;
 import com.synopsys.integration.alert.component.settings.descriptor.SettingsDescriptor;
 import com.synopsys.integration.alert.database.api.DefaultUserAccessor;
 import com.synopsys.integration.alert.database.api.SystemStatusUtility;
 import com.synopsys.integration.alert.database.system.SystemMessageUtility;
-import com.synopsys.integration.alert.provider.blackduck.BlackDuckProperties;
-import com.synopsys.integration.blackduck.configuration.BlackDuckServerConfig;
-import com.synopsys.integration.exception.IntegrationException;
-import com.synopsys.integration.log.IntLogger;
-import com.synopsys.integration.log.Slf4jIntLogger;
-import com.synopsys.integration.rest.proxy.ProxyInfo;
 
 @Component
 @Order(2)
 public class SystemValidator extends StartupComponent {
     private static final Logger logger = LoggerFactory.getLogger(SystemValidator.class);
-    private final AlertProperties alertProperties;
-    private final BlackDuckProperties blackDuckProperties;
+    private final List<ProviderValidator> providerValidators;
     private final EncryptionUtility encryptionUtility;
     private final SystemStatusUtility systemStatusUtility;
     private final SystemMessageUtility systemMessageUtility;
@@ -67,10 +57,9 @@ public class SystemValidator extends StartupComponent {
     private final ProxyManager proxyManager;
 
     @Autowired
-    public SystemValidator(final AlertProperties alertProperties, final BlackDuckProperties blackDuckProperties, final EncryptionUtility encryptionUtility, final SystemStatusUtility systemStatusUtility,
-        final SystemMessageUtility systemMessageUtility, final DefaultUserAccessor userAccessor, final ProxyManager proxyManager) {
-        this.alertProperties = alertProperties;
-        this.blackDuckProperties = blackDuckProperties;
+    public SystemValidator(final List<ProviderValidator> providerValidators, final EncryptionUtility encryptionUtility, final SystemStatusUtility systemStatusUtility, final SystemMessageUtility systemMessageUtility,
+        final DefaultUserAccessor userAccessor, final ProxyManager proxyManager) {
+        this.providerValidators = providerValidators;
         this.encryptionUtility = encryptionUtility;
         this.systemStatusUtility = systemStatusUtility;
         this.systemMessageUtility = systemMessageUtility;
@@ -177,70 +166,12 @@ public class SystemValidator extends StartupComponent {
     }
 
     public boolean validateProviders() {
-        final boolean valid;
+        boolean valid = true;
         logger.info("Validating configured providers: ");
-        valid = validateBlackDuckProvider();
+        for (final ProviderValidator providerValidator : providerValidators) {
+            valid = valid && providerValidator.validate();
+        }
         return valid;
     }
 
-    // TODO add this validation to provider descriptors so we can run this when it's defined (Or delete it entirely as we no longer require a BD setup)
-    public boolean validateBlackDuckProvider() {
-        logger.info("Validating BlackDuck Provider...");
-        systemMessageUtility.removeSystemMessagesByType(SystemMessageType.BLACKDUCK_PROVIDER_CONNECTIVITY);
-        systemMessageUtility.removeSystemMessagesByType(SystemMessageType.BLACKDUCK_PROVIDER_URL_MISSING);
-        systemMessageUtility.removeSystemMessagesByType(SystemMessageType.BLACKDUCK_PROVIDER_LOCALHOST);
-        try {
-            ProxyInfo proxyInfo;
-            try {
-                proxyInfo = proxyManager.createProxyInfo();
-            } catch (final IllegalArgumentException e) {
-                proxyInfo = ProxyInfo.NO_PROXY_INFO;
-            }
-
-            final Optional<String> blackDuckUrlOptional = blackDuckProperties.getBlackDuckUrl();
-            if (blackDuckUrlOptional.isEmpty()) {
-                logger.error("  -> BlackDuck Provider Invalid; cause: Black Duck URL missing...");
-                final String errorMessage = "BlackDuck Provider invalid: URL missing";
-                systemMessageUtility.addSystemMessage(errorMessage, SystemMessageSeverity.WARNING, SystemMessageType.BLACKDUCK_PROVIDER_URL_MISSING);
-            } else {
-                final String blackDuckUrlString = blackDuckUrlOptional.get();
-                final Boolean trustCertificate = BooleanUtils.toBoolean(alertProperties.getAlertTrustCertificate().orElse(false));
-                final Integer timeout = blackDuckProperties.getBlackDuckTimeout();
-                logger.debug("  -> BlackDuck Provider URL found validating: {}", blackDuckUrlString);
-                logger.debug("  -> BlackDuck Provider Trust Cert: {}", trustCertificate);
-                logger.debug("  -> BlackDuck Provider Timeout: {}", timeout);
-                final URL blackDuckUrl = new URL(blackDuckUrlString);
-                if ("localhost".equals(blackDuckUrl.getHost())) {
-                    logger.warn("  -> BlackDuck Provider Using localhost...");
-                    final String blackDuckWebServerHost = blackDuckProperties.getPublicBlackDuckWebserverHost().orElse("");
-                    logger.warn("  -> BlackDuck Provider Using localhost because PUBLIC_BLACKDUCK_WEBSERVER_HOST environment variable is set to {}", blackDuckWebServerHost);
-                    systemMessageUtility.addSystemMessage("BlackDuck Provider Using localhost", SystemMessageSeverity.WARNING, SystemMessageType.BLACKDUCK_PROVIDER_LOCALHOST);
-                }
-                final IntLogger intLogger = new Slf4jIntLogger(logger);
-                final Optional<BlackDuckServerConfig> blackDuckServerConfig = blackDuckProperties.createBlackDuckServerConfig(intLogger);
-                if (blackDuckServerConfig.isPresent()) {
-                    final Boolean canConnect = blackDuckServerConfig.get().canConnect(intLogger);
-                    if (canConnect) {
-                        logger.info("  -> BlackDuck Provider Valid!");
-                    } else {
-                        final String message = "Can not connect to the BlackDuck server with the current configuration.";
-                        connectivityWarning(systemMessageUtility, message);
-                    }
-                } else {
-                    final String message = "The BlackDuck configuration is not valid.";
-                    connectivityWarning(systemMessageUtility, message);
-                }
-            }
-        } catch (final MalformedURLException | IntegrationException | AlertRuntimeException ex) {
-            logger.error("  -> BlackDuck Provider Invalid; cause: {}", ex.getMessage());
-            logger.debug("  -> BlackDuck Provider Stack Trace: ", ex);
-            systemMessageUtility.addSystemMessage("BlackDuck Provider invalid: " + ex.getMessage(), SystemMessageSeverity.WARNING, SystemMessageType.BLACKDUCK_PROVIDER_CONNECTIVITY);
-        }
-        return true;
-    }
-
-    private void connectivityWarning(final SystemMessageUtility systemMessageUtility, final String message) {
-        logger.warn(message);
-        systemMessageUtility.addSystemMessage(message, SystemMessageSeverity.WARNING, SystemMessageType.BLACKDUCK_PROVIDER_CONNECTIVITY);
-    }
 }
