+++ src/test/java/com/synopsys/integration/alert/workflow/startup/SystemValidatorTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertTrue;
 
 import java.io.IOException;
 import java.util.HashMap;
+import java.util.List;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -16,6 +17,7 @@ import com.synopsys.integration.alert.common.security.EncryptionUtility;
 import com.synopsys.integration.alert.database.api.DefaultUserAccessor;
 import com.synopsys.integration.alert.database.api.SystemStatusUtility;
 import com.synopsys.integration.alert.database.system.SystemMessageUtility;
+import com.synopsys.integration.alert.provider.blackduck.BlackDuckValidator;
 import com.synopsys.integration.alert.provider.blackduck.TestBlackDuckProperties;
 import com.synopsys.integration.alert.util.OutputLogger;
 import com.synopsys.integration.alert.util.TestAlertProperties;
@@ -52,7 +54,7 @@ public class SystemValidatorTest {
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
         final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
         final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
+        final SystemValidator systemValidator = new SystemValidator(List.of(), encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
         final SystemValidator spiedSystemValidator = Mockito.spy(systemValidator);
 
         spiedSystemValidator.validate();
@@ -61,14 +63,12 @@ public class SystemValidatorTest {
 
     @Test
     public void testValidateEncryptionProperties() throws IOException {
-        final TestAlertProperties testAlertProperties = new TestAlertProperties();
-        final TestBlackDuckProperties testGlobalProperties = new TestBlackDuckProperties(testAlertProperties);
         final EncryptionUtility encryptionUtility = Mockito.mock(EncryptionUtility.class);
         final SystemStatusUtility systemStatusUtility = Mockito.mock(SystemStatusUtility.class);
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
         final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
         final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
+        final SystemValidator systemValidator = new SystemValidator(List.of(), encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
         systemValidator.validateEncryptionProperties(new HashMap<>());
         Mockito.verify(encryptionUtility).isInitialized();
         assertTrue(outputLogger.isLineContainingText("Encryption utilities: Not Initialized"));
@@ -76,15 +76,13 @@ public class SystemValidatorTest {
 
     @Test
     public void testValidateEncryptionPropertiesSuccess() throws IOException {
-        final TestAlertProperties testAlertProperties = new TestAlertProperties();
-        final TestBlackDuckProperties testGlobalProperties = new TestBlackDuckProperties(testAlertProperties);
         final EncryptionUtility encryptionUtility = Mockito.mock(EncryptionUtility.class);
         final SystemStatusUtility systemStatusUtility = Mockito.mock(SystemStatusUtility.class);
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
         Mockito.when(encryptionUtility.isInitialized()).thenReturn(true);
         final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
         final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
+        final SystemValidator systemValidator = new SystemValidator(List.of(), encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
         systemValidator.validateEncryptionProperties(new HashMap<>());
         Mockito.verify(encryptionUtility).isInitialized();
         assertTrue(outputLogger.isLineContainingText("Encryption utilities: Initialized"));
@@ -101,7 +99,7 @@ public class SystemValidatorTest {
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
         final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
         final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
+        final SystemValidator systemValidator = new SystemValidator(List.of(), encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
         systemValidator.validateProviders();
         assertTrue(outputLogger.isLineContainingText("Validating configured providers: "));
     }
@@ -110,14 +108,9 @@ public class SystemValidatorTest {
     public void testvalidateBlackDuckProviderNullURL() throws IOException {
         final TestAlertProperties testAlertProperties = new TestAlertProperties();
         final TestBlackDuckProperties testGlobalProperties = new TestBlackDuckProperties(testAlertProperties);
-        final EncryptionUtility encryptionUtility = Mockito.mock(EncryptionUtility.class);
-        final SystemStatusUtility systemStatusUtility = Mockito.mock(SystemStatusUtility.class);
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
-        final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
-        final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
-        testGlobalProperties.setBlackDuckUrl(null);
-        systemValidator.validateBlackDuckProvider();
+        final BlackDuckValidator blackDuckValidator = new BlackDuckValidator(testAlertProperties, testGlobalProperties, systemMessageUtility);
+        blackDuckValidator.validate();
         assertTrue(outputLogger.isLineContainingText("Validating BlackDuck Provider..."));
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Invalid; cause: Black Duck URL missing..."));
     }
@@ -126,14 +119,10 @@ public class SystemValidatorTest {
     public void testvalidateBlackDuckProviderLocalhostURL() throws IOException {
         final TestAlertProperties testAlertProperties = new TestAlertProperties();
         final TestBlackDuckProperties testGlobalProperties = new TestBlackDuckProperties(testAlertProperties);
-        final EncryptionUtility encryptionUtility = Mockito.mock(EncryptionUtility.class);
-        final SystemStatusUtility systemStatusUtility = Mockito.mock(SystemStatusUtility.class);
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
-        final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
-        final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
+        final BlackDuckValidator blackDuckValidator = new BlackDuckValidator(testAlertProperties, testGlobalProperties, systemMessageUtility);
         testGlobalProperties.setBlackDuckUrl("https://localhost:443");
-        systemValidator.validateBlackDuckProvider();
+        blackDuckValidator.validate();
         assertTrue(outputLogger.isLineContainingText("Validating BlackDuck Provider..."));
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Using localhost..."));
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Using localhost because PUBLIC_BLACKDUCK_WEBSERVER_HOST environment variable is set to"));
@@ -145,13 +134,9 @@ public class SystemValidatorTest {
         final TestBlackDuckProperties testGlobalProperties = new TestBlackDuckProperties(testAlertProperties);
         final TestBlackDuckProperties spiedGlobalProperties = Mockito.spy(testGlobalProperties);
         spiedGlobalProperties.setBlackDuckUrl("https://localhost:443");
-        final EncryptionUtility encryptionUtility = Mockito.mock(EncryptionUtility.class);
-        final SystemStatusUtility systemStatusUtility = Mockito.mock(SystemStatusUtility.class);
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
-        final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
-        final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, spiedGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
-        systemValidator.validateBlackDuckProvider();
+        final BlackDuckValidator blackDuckValidator = new BlackDuckValidator(testAlertProperties, testGlobalProperties, systemMessageUtility);
+        blackDuckValidator.validate();
         assertTrue(outputLogger.isLineContainingText("Validating BlackDuck Provider..."));
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Using localhost..."));
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Using localhost because PUBLIC_BLACKDUCK_WEBSERVER_HOST environment variable is set to"));
@@ -161,14 +146,10 @@ public class SystemValidatorTest {
     public void testValidateHubInvalidProvider() throws IOException {
         final TestAlertProperties testAlertProperties = new TestAlertProperties();
         final TestBlackDuckProperties testGlobalProperties = new TestBlackDuckProperties(testAlertProperties);
-        final EncryptionUtility encryptionUtility = Mockito.mock(EncryptionUtility.class);
-        final SystemStatusUtility systemStatusUtility = Mockito.mock(SystemStatusUtility.class);
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
-        final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
-        final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
+        final BlackDuckValidator blackDuckValidator = new BlackDuckValidator(testAlertProperties, testGlobalProperties, systemMessageUtility);
         testGlobalProperties.setBlackDuckUrl("https://localhost:443");
-        systemValidator.validateBlackDuckProvider();
+        blackDuckValidator.validate();
         assertTrue(outputLogger.isLineContainingText("Validating BlackDuck Provider..."));
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Using localhost..."));
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Invalid; cause:"));
@@ -179,13 +160,9 @@ public class SystemValidatorTest {
     public void testValidateHubValidProvider() throws IOException {
         final TestAlertProperties testAlertProperties = new TestAlertProperties();
         final TestBlackDuckProperties testGlobalProperties = new TestBlackDuckProperties(testAlertProperties);
-        final EncryptionUtility encryptionUtility = Mockito.mock(EncryptionUtility.class);
-        final SystemStatusUtility systemStatusUtility = Mockito.mock(SystemStatusUtility.class);
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
-        final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
-        final ProxyManager proxyManager = Mockito.mock(ProxyManager.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
-        systemValidator.validateBlackDuckProvider();
+        final BlackDuckValidator blackDuckValidator = new BlackDuckValidator(testAlertProperties, testGlobalProperties, systemMessageUtility);
+        blackDuckValidator.validate();
         assertTrue(outputLogger.isLineContainingText("Validating BlackDuck Provider..."));
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Valid!"));
     }
@@ -213,12 +190,9 @@ public class SystemValidatorTest {
         testGlobalProperties.setBlackDuckUrl("Black Duck URL");
         testGlobalProperties.setBlackDuckApiKey("Black Duck API Token");
         testGlobalProperties.setBlackDuckTimeout(1);
-        final EncryptionUtility encryptionUtility = Mockito.mock(EncryptionUtility.class);
-        final SystemStatusUtility systemStatusUtility = Mockito.mock(SystemStatusUtility.class);
         final SystemMessageUtility systemMessageUtility = Mockito.mock(SystemMessageUtility.class);
-        final DefaultUserAccessor userAccessor = Mockito.mock(DefaultUserAccessor.class);
-        final SystemValidator systemValidator = new SystemValidator(testAlertProperties, testGlobalProperties, encryptionUtility, systemStatusUtility, systemMessageUtility, userAccessor, proxyManager);
-        systemValidator.validateBlackDuckProvider();
+        final BlackDuckValidator blackDuckValidator = new BlackDuckValidator(testAlertProperties, testGlobalProperties, systemMessageUtility);
+        blackDuckValidator.validate();
         assertTrue(outputLogger.isLineContainingText("BlackDuck Provider Invalid; cause:"));
     }
 }
